+++ extensions/git/src/components/actions/RemoteActions.tsx
@@ -16,6 +16,7 @@ export function RemoteFetchAction(context: RepositoryContext & { remotesHosts?: 
       context.branches.revalidate();
       context.commits.revalidate();
       context.status.revalidate();
+      context.tags.revalidate();
     } catch {
       // Git error is already shown by GitManager
     }
@@ -249,9 +250,12 @@ export namespace RemoteWebPageAction {
     }
   }
 
-  export function Base({ remote }: { remote: Remote }) {
+  export function Base({ remote, showTitle = false }: { remote: Remote; showTitle?: boolean }) {
     return (
-      <ActionPanel.Section>
+      <ActionPanel.Section
+        key={`remote-web-page-other-${remote.name}`}
+        title={showTitle ? remote.displayName : undefined}
+      >
         {remote.webPages.other().map((page, index) => (
           <Action.OpenInBrowser key={`remote-web-page-other-${remote.name}-${index}`} {...page} />
         ))}