+++ extensions/paste-to-markdown/src/joplin__turndown-plugin-gfm.d.ts
@@ -1,0 +1,8 @@
+declare module "@joplin/turndown-plugin-gfm" {
+  import TurndownService from "turndown";
+
+  export function gfm(service: TurndownService): void;
+  export function tables(service: TurndownService): void;
+  export function strikethrough(service: TurndownService): void;
+  export function taskListItems(service: TurndownService): void;
+}